#refactor: feito ajustes estruturais, clean code e ajustado configs da API
