+++ src/Backend/BrazilianAddresses.Api/appsettings.Development.json
@@ -4,5 +4,8 @@
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
+  },
+  "ConnectionStrings": {
+    "Connection": "Server=server-name;Initial Catalog=challenge-addresses;Persist Security Info=False;User ID=user-name;Password=user-password;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=300;"
   }
 }
