+++ src/Backend/BrazilianAddresses.Application/BusinessRules/IBGEBusinessRule/Interfaces/ICreateIBGE.cs
@@ -1,7 +1,7 @@
 ﻿using BrazilianAddresses.Communication.Requests;
 using BrazilianAddresses.Communication.Responses;
 
-namespace BrazilianAddresses.Application.BusinessRules.IBGEBusinessRule
+namespace BrazilianAddresses.Application.BusinessRules.IBGEBusinessRule.Interfaces
 {
     public interface ICreateIBGE
     {
