+++ src/Backend/BrazilianAddresses.Domain/Enum/StateEnum.cs
@@ -59,5 +59,4 @@ namespace BrazilianAddresses.Domain.Enum
         [Description("Tocantins")]
         TO = 27
     }
-
 }
